pyproject.toml
[project]
name = "2048mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.0",
    "flask>=3.1.0",
    "nuitka>=2.6.9",
    "pillow>=11.2.1",
    "requests>=2.32.3",
]
